Extract IndividualRouteExport.Export
